# Conflicts:
#	src/main/java/com/corporation/controller/PostController.java
#	src/main/java/com/corporation/model/Post.java
#	src/main/java/com/corporation/repository/PostRepository.java
#	src/main/java/com/corporation/service/PostService.java
#	src/main/resources/db/changelog/db.changelog-master.yaml
#	src/main/resources/db/changelog/init-post.sql
#	src/test/java/com/corporation/service/PostServiceTest.java

+++ src/test/java/com/corporation/service/UserServiceTest.java
@@ -27,8 +27,7 @@ public class UserServiceTest {
 
     @Test
     public void shouldReturnUserById() {
-
-        int desiredId = 1;
+        long desiredId = 1;
         String nickname = "boba";
         String email = "boba@boba.com";
         String password = "1234";
@@ -57,8 +56,7 @@ public class UserServiceTest {
 
     @Test
     public void shouldThrowUserNotFoundException() {
-
-        int desiredId = 100;
+        long desiredId = 100;
 
         Mockito.when(userRepository.findById(desiredId))
                 .thenReturn(Optional.empty());
